feat: fixed and update register form. refactor: split big component into smaller ones
