+++ app/components/auth/register.scss
@@ -1,7 +1,16 @@
+@import '../../styles/mixins.scss';
+@import '../../styles/variables.scss';
+@import '../../styles/function.scss';
+@import '../../styles/colors.scss';
+
+
 .wrapper_register-form {
 	width: 100%;
 	height: 100dvh;
-	background: linear-gradient(180deg, #627254, #050403);
+	background-image: url('../../assets/images/form-\ image.png');
+	background-position: center;
+	background-repeat: no-repeat;
+	background-size: cover;
 	border-radius: 50px;
 	display: flex;
 	flex-direction: column;
@@ -12,27 +21,27 @@
 	& .register_form-container {
 		display: grid;
 		grid-template-columns: 1fr;
-		grid-template-rows: auto 1fr auto;
+		grid-template-rows: auto auto 1fr auto;
 		gap: 0px;
+		background-color: color(white, light);
+
+		& .register_icon {
+			margin: 16px auto;
+		}
 
 		& .register_form-head {
 			display: flex;
 			flex-direction: column;
 			justify-content: center;
 			align-items: center;
-			margin-bottom: 26px;
 
-			& .signin_icon {
-				width: 32.5px;
-				height: 30px;
-				margin-bottom: 20px;
-			}
 			& .register_title {
-				font-size: 24px;
-				font-weight: 700;
+				@include heading-sm();
+				color: color(green, darker);
 				line-height: 1;
 				letter-spacing: 0;
 				text-align: center;
+				margin-bottom: 20px;
 			}
 		}
 
@@ -53,16 +62,16 @@
 					height: 45px;
 					padding-left: 17px;
 					border-radius: 15px;
-					border: 1px solid #ffffff;
+					border: 1px solid color( green, darker);
 					outline: none;
 					background-color: transparent;
 
 					&::placeholder {
-						color: #ffffff;
-						font-size: 14px;
+						color: color(white, darker);
+						@include body-sm();
 						line-height: 1;
 						letter-spacing: 0;
-						font-weight: regular;
+
 					}
 				}
 				& .password_box {
@@ -95,13 +104,6 @@
 				}
 			}
 
-			// & .checkbox_box {
-			// 	display: flex;
-			// 	flex-direction: column;
-			// 	margin-top: 20px;
-			// 	margin-inline: 25px;
-			// 	display: flex;
-
 				& .checkbox_container {
 					display: flex;
 					flex-direction: row;
@@ -115,8 +117,13 @@
 						border-color: #ffffff;
 						background-color: transparent;
 					}
+
+					& span {
+						@include body-sm();
+						color: color(green, darker);
+						line-height: 1;
+					}
 				}
-			// }
 			& .submit_button {
 				display: inline-flex;
 				justify-content: center;
@@ -124,10 +131,11 @@
 				text-align: center;
 				height: 26px;
 				margin-top: 32px;
+				padding: 16px 0;
 				margin-inline: 25px;
-				background-color: #ffffff;
-				border-radius: 10px;
-				color: #000000;
+				background-color: color(green, darker);
+				border-radius: 15px;
+				color: color(white, light);
 				font-size: 15px;
 				font-weight: 600;
 				cursor: pointer;
@@ -143,12 +151,14 @@
 
 		& .login_redirect {
 			text-align: center;
-			font-size: 10px;
 			line-height: 1;
-			font-weight: 500;
+			@include body-sm();
+			color: color(green, darker);
+			margin-bottom: 16px;
 			& .login_redirect-link {
+				margin-left:8px;
 				cursor: pointer;
-				color: #00ff66;
+				color: color(green, normal-hover)
 			}
 		}
 		& .error_message {
@@ -181,13 +191,6 @@
 @media screen and (min-width: 992px) and (max-width: 1200px) {
 	.wrapper_register-form {
 		max-width: 430px;
-		& .register_form-container {
-			& .register_form-head {
-				& .register_title {
-					font-size: 36px;
-				}
-			}
-		}
 	}
 }
 @media screen and (min-width: 1201px) and (max-width: 1440px) {
