+++ app/components/main/benefits-section/styles.scss
@@ -1,0 +1,65 @@
+@import '../../../styles/colors.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/function.scss';
+
+.benefits_section {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+
+		& .benefits_section-header {
+			& .benefits_section-title {
+				@include body-lg();
+				color: color(green, darker);
+				text-align: center;
+				margin-bottom: 14px;
+			}
+			& .benefits_section-description {
+				@include body-sm();
+				color: color(green, dark-hover);
+				text-align: center;
+				
+			}
+		}
+
+		& .benefits_cards {
+			& .benefits_card {
+				margin: 8px 0px;
+				& .benefits_card-container {
+					background-color: color(green, light);
+					border-radius: 10px;
+					padding: 28.5px 24px;
+
+					& .granteed_quality,
+					& .free_delivery,
+					& .expert_consultation,
+					& .eco_packaging {
+						display: grid;
+						grid-template-columns: auto 1fr;
+						align-items: center;
+
+						margin-bottom: 12px;
+
+						& h5 {
+							@include body-md();
+							color: color(green, darker);
+							line-height: 1.2;
+						}
+						& .granteed_quality-icon,
+						& .free_delivery-icon,
+						& .expert_consultation-icon,
+						& .eco_packaging-icon {
+							margin-right: 15px;
+						}
+					}
+
+					& span {
+						@include body-sm();
+						color: color(green, darker);
+						margin-top: auto;
+					}
+				}
+			}
+		}
+	}
