+++ app/components/main/catalog-section/styles.scss
@@ -1,0 +1,115 @@
+@import '../../../styles/colors.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/function.scss';
+
+.catalog_offers_section {
+		margin: 56px 16px;
+		& .catalog_offers-container {
+			& .slick-slider {
+				display: grid;
+				grid-template-columns: repeat(2, 1fr);
+				grid-template-rows: 1fr auto;
+			}
+
+			& .slider-controls {
+				display: flex;
+				justify-content: center;
+				gap: 16px;
+			}
+
+			.slick-slide > div {
+				height: 100%;
+				overflow: hidden;
+			}
+
+
+			& .slick-prev {
+				grid-column: 1/2;
+				grid-row: 2/3;
+				background-image: url(../../../assets/left-arrow-icon.svg);
+			}
+
+			& .slick-next {
+				grid-column: 2/3;
+				grid-row: 2/3;
+				background-image: url(../../../assets/right-arrow-icon.svg);
+			}
+
+			& .slick-prev,
+			& .slick-next {
+				margin: 0 auto;
+				margin-top: 5px;
+				border: none;
+				outline: none;
+				height: 15px;
+				width: 50px;
+				background-repeat: no-repeat;
+				background-position: center;
+				background-color: transparent;
+			}
+
+			& .slick-list {
+				display: flex;
+				overflow: hidden;
+				grid-column: 1/3;
+				grid-row: 1/2;
+				margin: 0 15px;
+			}
+
+			& .slick-track {
+				display: flex !important;
+			}
+
+			& .slick-slide {
+				display: flex !important;
+				justify-content: center;
+			}
+
+			& .catalog_offers-title {
+				@include body-lg();
+				text-align: center;
+				color: #0a2308;
+			}
+			& .catalog_offers-description {
+				margin-top: 16px;
+				@include body-sm();
+				color: color(green, dark-hover);
+				text-align: center;
+				word-wrap: normal;
+			}
+			& .catalog_offers-cards {
+				margin-top: 32px;
+
+				& .catalog_offers-card {
+					& .catalog_offers-card-content {
+						display: flex;
+						flex-direction: row;
+						justify-content: space-between;
+						align-items: center;
+						margin-top: 2px;
+
+						& .catalog_offers-card-title {
+							@include body-md();
+							color: color(green, dark-active);
+						}
+
+						& .catalog_offers-card-link {
+							color: color(green, dark-active);
+							@include body-lg();
+							text-decoration: none;
+						}
+					}
+				}
+			}
+		}
+		& .catalog_offers-link {
+			margin-top: 12px;
+			width: 100%;
+			text-align: center;
+			& .catalog_offers-link-text {
+				@include body-lg();
+				color: color(green, darker);
+			}
+		}
+	}
