+++ app/components/main/faq-section/styles.scss
@@ -1,0 +1,53 @@
+@import '../../../styles/colors.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/function.scss';
+
+.faq_section {
+		margin-top: 56px;
+		& .faq_section-title {
+			@include body-lg();
+			color: #020201;
+			text-align: center;
+			margin-bottom: 14px;
+		}
+		& .faq_section-description {
+			@include body-md();
+			color: #6e6e69;
+			text-align: center;
+		}
+
+		& .faq_section-container {
+			margin-top: 24px;
+			& .faq_card-container {
+				margin: 6px 0px;
+
+				& .faq_card {
+					display: grid;
+					grid-template-columns: 1fr auto;
+					place-content: center;
+					background-color: color(green, light);
+					border-radius: 10px;
+					padding: 28.5px 24px;
+
+					& .faq_card-question {
+						@include body-sm();
+						color: color(green, darker);
+						align-self: center;
+					}
+
+					& .icon-box {
+						justify-self: end;
+						display: flex;
+						justify-content: center;
+						align-items: center;
+						width: 30px;
+						height: 30px;
+						background-color: color(white, light);
+						border-radius: 26px;
+						margin-left: 10px;
+					}
+				}
+			}
+		}
+	}
