+++ app/components/main/hero-section/styles.scss
@@ -1,0 +1,48 @@
+@import '../../../styles/colors.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/mixins.scss';
+@import '../../../styles/function.scss';
+
+.hero_section {
+		background-image: url('../../../assets/images/hero\ image.png');
+		background-repeat: no-repeat;
+		background-size: cover;
+		background-position: center;
+		width: 100%;
+		height: 700px;
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		justify-content: start;
+		& .hero_section-header {
+			margin-top: 96px;
+			max-width: 858px;
+			margin-bottom: 19px;
+			& .hero_section-title {
+				@include heading-sm();
+				margin: 0 19px;
+				text-align: center;
+				color: #fffffa;
+			}
+			& .hero_section-description {
+				@include body-sm();
+				margin: 0 19px;
+				margin-top: 8px;
+				text-align: center;
+				color: #fffffa;
+			}
+		}
+		& .to_catalog {
+			& .to_catalog-button {
+				margin: 78px 19px 0;
+				min-width: 270px;
+				height: 52px;
+				border-radius: 10px;
+				border: none;
+				outline: none;
+				@include body-lg();
+				color: #000000;
+				background-color: color(white, light);
+			}
+		}
+	}
