+++ app/hooks/useRegistartionForm.ts
@@ -1,7 +1,9 @@
-
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { registrationSchema, RegistrationFormData } from '@features/auth/auth.schema';
+import {
+	registrationSchema,
+	RegistrationFormData,
+} from '@features/auth/auth.schema';
 
 export const useRegistrationForm = () =>
 	useForm<RegistrationFormData>({
